@@ -25,10 +25,10 @@ task clean(type: Delete) {
 }
 
 ext {
-    minSdkVersion = 9
-    targetSdkVersion = 26
-    compileSdkVersion = 26
-    buildToolsVersion = '27.0.3'
+    minSdkVersion = 14
+    targetSdkVersion = 30
+    compileSdkVersion = 30
+    buildToolsVersion = '30.0.2'
     sourceCompatibilityVersion = JavaVersion.VERSION_1_8
     targetCompatibilityVersion = JavaVersion.VERSION_1_8
 }
@@ -40,9 +40,9 @@ ext.deps = [
         robolectric       : 'org.robolectric:robolectric:3.0',
 
         // dependencies
-        appcompatV7       : 'com.android.support:appcompat-v7:25.4.0',
-        androidannotations: 'com.android.support:support-annotations:25.4.0',
-        design            : 'com.android.support:design:25.4.0',
+        appcompatV7       : 'androidx.appcompat:appcompat:1.0.0',
+        androidannotations: 'androidx.annotation:annotation:1.0.0',
+        design            : 'com.google.android.material:material:1.0.0',
 
         rxandroid         : 'io.reactivex:rxandroid:1.2.1',
         rxjava            : 'io.reactivex:rxjava:1.3.4',
